feat(api-uc-imp-upd): fix tests and improve process refs: #359
