+++ api/src/core/adapters/dbApi/kysely/kysely.utils.ts
@@ -42,8 +42,23 @@ export const stripNullOrUndefinedValues = <T extends Record<string, unknown>>(
 ): { [K in keyof T]: null extends T[K] ? Exclude<T[K], null> | undefined : T[K] } =>
     Object.fromEntries(Object.entries(obj).filter(([_, value]) => value !== null && value !== undefined)) as any;
 
+export const transformNullToUndefined = <T extends Record<string, any>>(obj: T): TransformObjToNullToUndefined<T> => {
+    const transformedObj = { ...obj } as { [K in keyof T]: T[K] };
+
+    for (const key in obj) {
+        if (transformedObj[key] === null) {
+            transformedObj[key] = undefined as any;
+        }
+    }
+    return transformedObj;
+};
+
 // Utility type to convert null to undefined
 type NullToUndefined<T> = T extends null ? undefined : T;
+
+export type TransformObjToNullToUndefined<O> = {
+    [K in keyof O]: NullToUndefined<O[K]>;
+};
 
 // Utility type to remove JSONColumnType
 type RemoveJSONColumnType<T> = T extends JSONColumnType<infer U> ? U : T;
@@ -52,6 +67,23 @@ type RemoveJSONColumnType<T> = T extends JSONColumnType<infer U> ? U : T;
 type RemoveGeneratedType<T> = T extends Generated<infer U> ? U : T;
 
 // Transform the SoftwaresTable type
-export type TransformRepoToRaw<R> = {
+export type TransformRepoToRowOutput<R> = {
+    [K in keyof R]: RemoveGeneratedType<RemoveJSONColumnType<R[K]>>;
+};
+
+// Transform the SoftwaresTable type
+export type TransformRepoToCleanedRow<R> = {
     [K in keyof R]: RemoveGeneratedType<NullToUndefined<RemoveJSONColumnType<R[K]>>>;
+};
+
+export const parseBigIntToNumber = <T extends Object>(obj: T, keys: (keyof T)[]) => {
+    const transformedObj = { ...obj };
+
+    keys.forEach(key => {
+        if (key in obj && typeof obj[key] === "string") {
+            transformedObj[key] = new Date(+transformedObj[key]).getTime() as T[keyof T];
+        }
+    });
+
+    return transformedObj;
 };
